Added Upgrades Menu
Added all current assets for the upgrades menu and scripted to allow the addition of more later.
All relevant hooks to gameplay variables are in and the menu updates based on what the player has done / will be doing.
Fixed a bug where the AudioInterface was called before being initialised.
Fixed all animations to have parameters being called by scripts.