Codebase Refactor and Start of Throw the Robot
Moved much of the data being stored in PlayerPrefs into a GameManager class, which handles holding this and saving it on level or game unload, as well as loading it at the beginning.
Saves and settings are now serialised to a file.
Level manager base class and variants for each level make level sequencing a more pain free process.
Many misc fixes to small bugs like the twin jittering and music getting confused.
Story sequencing is now far more deliberate and less maths based, such that levels and cutscenes can be made to open / occur any time.
Early versions of throw mechanics, level builder and player controls for Throw the Robot added.
Upgrades UI totally reworked a to allow for multiple tiers of upgrades and easier set up of upgrades.