Freeze Ability and New Ripples
New Ripple animations in for map trash.
Added assets and behaviours to Scrap Grabber for the freeze ability.
Changed some time related variables on the claw to make frozen time smoother.
Added upgrade code to the claw and gave this the ability to activate both the extension and freeze.
Edited MinigameUI to get rid of a bug where it would play the game-over sound new every frame.
Added new ripple sprites and set these up in their respective animations.