Added Upgrade HUD, Electric Eel and Some Fixes
Electric Eel sprites animations, behaviours, and prefab added.
HUD element for boost and time freeze added, along with all sprites animations and behaviours.  This can be recycled for other active upgrades.
Fixed Brainy's request system coming apart due to smooth animation of the panels.
LevelManager_Map now respawns trash on each load rather than only the first one.
Random trash script edited to make sure it can find the levelmanager in the case that it needs respawning on collision.
LightsOut coroutine added to LevelManager_ScrapGrabber, takes the eel that was hit as an overload, though this defaults to null in case something else triggers it in future.
Made spawner dispose of seeds on disable to avoid memory leaks.
Added animation code for the boost HUD to the Player script.
Updated Anglerfish spritesheet to one containing the sprites needed for fleeing.