Upgrade Tutorial System and some QOL Changes
Added assets, animations, behaviours, and prefabs for the upgrades tutorial system.
Edited the PauseMenu settings for each level to only show the correct levels in upgrades tutorial book.
Added assets, animations, and behaviours for the upgrades preview window.
Added functionality to show mouse when clicked on levels where the mouse is hidden.